Moving MatchCard into shared UI.
