+++ android/design-system/src/androidTest/kotlin/com/adammcneilly/pocketleague/android/designsystem/components/MatchCardTest.kt
@@ -8,6 +8,7 @@ import androidx.compose.ui.test.performClick
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.test.TestDisplayModel
 import com.adammcneilly.pocketleague.shared.design.system.theme.PocketLeagueTheme
+import com.adammcneilly.pocketleague.shared.ui.match.MatchCard
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Rule
