+++ shared/ui/build.gradle.kts
@@ -15,6 +15,7 @@ kotlin {
                 implementation(project(":shared:design-system"))
                 implementation(compose.foundation)
                 implementation(compose.material3)
+                implementation(compose.materialIconsExtended)
                 implementation(compose.preview)
                 implementation(compose.runtime)
             }
