+++ shared/ui/src/test/snapshots/images/com.adammcneilly.pocketleague.shared.ui.match_MatchCardPaparazziTest_renderOrangeTeamWinner[useDarkTheme=true].png
